pyproject.toml
[tool.poetry]
name = "messenger_api"
version = "0.1.0"
description = ""
authors = ["Etozheigor <itshenko.io@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djoser = "^2.2.0"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
drf-yasg = "^1.21.7"
pillow = "^10.0.0"
python-dotenv = "^1.0.0"
django-debug-toolbar = "^4.2.0"
djangorestframework-simplejwt = "^5.2.2"
flake8 = "^6.1.0"
flake8-isort = "^6.0.0"
black = "^23.7.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
django-cleanup = "^8.0.0"
django-filter = "^23.2"
django-phonenumber-field = "^7.1.0"
drf-extra-fields = "^3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
